+++ templates/helpers/typescript-formatter.js
@@ -54,6 +54,8 @@ function formatImportStatements(importTypes, {
   importTypes.forEach(type => {
     if (type === 'Response') {
       importStatements.push('import { Response } from \'node-fetch\';');
+    } else if (type === 'ReadStream') {
+      importStatements.push('import { ReadStream } from \'fs\';');
     } else if (type === 'Collection') {
       importStatements.push(`import { Collection } from '${isModelToModelImport ? '..' : '.'}/collection';`);
     } else {